+++ src/frontend/src/views/Data/components/TreesView/components/TreeCreateHelpLink/styles.ts
@@ -6,7 +6,7 @@ import {
   getFontWeights,
   getSpaces,
 } from "czifui";
-import { NewTabLink } from "../../../NewTabLink";
+import { NewTabLink } from "../../../../../../common/components/library/NewTabLink";
 
 export const StyledDiv = styled.div`
   ${(props: CommonThemeProps) => {
